+++ x-pack/platform/plugins/shared/index_management/__jest__/client_integration/home/enrich_policies.test.tsx
@@ -42,13 +42,11 @@ describe('Enrich policies tab', () => {
       setDelayResponse(true);
       httpRequestsMockHelpers.setLoadEnrichPoliciesResponse([]);
 
-      const renderPromise = renderHome(httpSetup, {
+      renderHome(httpSetup, {
         initialEntries: ['/enrich_policies'],
       });
 
       expect(await screen.findByTestId('sectionLoading')).toBeInTheDocument();
-
-      await renderPromise;
     });
 
     test('displays a error prompt', async () => {
