+++ x-pack/platform/plugins/shared/index_management/__jest__/client_integration/index_template_wizard/template_clone.test.tsx
@@ -34,7 +34,7 @@ describe('<TemplateClone />', () => {
   });
 
   // RTL test - migrated
-  describe('page title (RTL)', () => {
+  describe('page title', () => {
     beforeEach(async () => {
       renderTemplateClone(httpSetup);
       await screen.findByTestId('pageTitle');
