+++ x-pack/platform/plugins/shared/index_management/public/application/components/component_templates/__jest__/client_integration/component_template_edit.test.tsx
@@ -29,8 +29,6 @@ describe('<ComponentTemplateEdit />', () => {
     jest.spyOn(breadcrumbService, 'setBreadcrumbs');
   });
 
-  afterAll(() => {});
-
   beforeEach(() => {
     jest.clearAllMocks();
     const env = setupEnvironment();
